+++ applications/backend/src/declarations/passport.d.ts
@@ -1,0 +1,13 @@
+import { UserEntity } from '../entities/user.entity';
+
+declare global {
+  namespace Express {
+    interface User extends UserEntity {}
+  }
+}
+
+declare module 'express-session' {
+  interface SessionData {
+    passport: { user: Express.User };
+  }
+}
