+++ applications/backend/src/types/socket.type.ts
@@ -1,12 +1,6 @@
 import { Socket as S } from 'socket.io';
-import session from 'express-session';
-import { UserEntity } from '../entities/user.entity';
+import { Express } from 'express';
 
 export type Socket = S & {
-  request: {
-    user: UserEntity;
-    session: session.Session & Partial<session.SessionData>;
-    sessionID: string;
-    sessionStore: Express.SessionStore;
-  };
+  request: Express.Request;
 };
